Experimental Code
Two modes: 1) Quantised and 0) UnQuantised.
In Quantised, the note that is coming out of CV Out is accented.
In UnQuatised mode, CV In goes straight through to CV Out, with the LEDs lighting to show the level if CV In.

If any LEDs are pressed, the module is in Quantised mode. With all LEDs out, it's in UnQuantised mode.